python/requirements.txt
#lokales Modul
#-e ./src/rulebot

PyMuPDF==1.24.11
openai>=1.33
transformers==4.48.0
datasets==3.2.0
evaluate==0.4.3
faiss-cpu==1.9.0.post1
numpy==1.26.3
scikit-learn==1.6.1
spacy==3.8.3
sentence_transformers==3.4.1
fastapi==0.115.8
pydantic==2.9.2
starlette==0.45.3
uvicorn==0.34.0

accelerate>=0.26.0
#--index-url https://download.pytorch.org/whl/cu124
torch==2.5.1+cu124
#--index-url https://download.pytorch.org/whl/cu124
torchvision==0.20.1+cu124
#--index-url https://download.pytorch.org/whl/cu124
torchaudio==2.5.1+cu124
#--index-url https://download.pytorch.org/whl/cu124